fix: update to ansible-runner to support python10+
ansible-runner came out 1 year after the machine-stats and the library
matched how we wanted to use ansible, as a way to
interact with ansible through a custom cli. So, I've added
ansible-runner to be our primary interface to ansible and plan on
cleaning up the code in __init__ to more closely match our new usage.