feat: Update support for Python versions > 3.10 (#286)
* chore: update ansible versions

* chore: update to ansible 2.12

* fix: update to ansible-runner to support python10+

ansible-runner came out 1 year after the machine-stats and the library
matched how we wanted to use ansible, as a way to
interact with ansible through a custom cli. So, I've added
ansible-runner to be our primary interface to ansible and plan on
cleaning up the code in __init__ to more closely match our new usage.

* chore: extract config to it's own library

* chore: add minimal ansible requirements

* chore: add default ServerAliveInterval to deal with Ansible freezing

* fix: name configuration file being used

* chore: update requirements .txt

* chore: update to python314

* chore: update setup.py

* feat: support a version flag argument in the machine_stats cli

* chore: add deprecation notice for virtstats

* chore: fix tests